Move constants, rename variable, re-order code
