+++ measures/views.py
@@ -38,11 +38,11 @@ from common.views import TrackedModelDetailView
 from geo_areas.models import GeographicalArea
 from geo_areas.utils import get_all_members_of_geo_groups
 from measures import forms
+from measures.constants import MEASURE_CONDITIONS_FORMSET_PREFIX
 from measures.constants import START
 from measures.constants import MeasureEditSteps
 from measures.filters import MeasureFilter
 from measures.filters import MeasureTypeFilterBackend
-from measures.forms import MEASURE_CONDITIONS_FORMSET_PREFIX
 from measures.models import FootnoteAssociationMeasure
 from measures.models import Measure
 from measures.models import MeasureActionPair
@@ -972,20 +972,20 @@ class MeasureUpdateBase(
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        initial = self.request.session.get(
+        context["no_form_tags"] = FormHelper()
+        context["no_form_tags"].form_tag = False
+
+        formset_footnotes = self.request.session.get(
             f"formset_initial_{self.kwargs.get('sid')}",
             [],
         )
         footnotes_formset = forms.MeasureUpdateFootnotesFormSet()
-        footnotes_formset.initial = initial
+        footnotes_formset.initial = formset_footnotes
         footnotes_formset.form_kwargs = {"path": self.request.path}
         context["footnotes_formset"] = footnotes_formset
-        context["no_form_tags"] = FormHelper()
-        context["no_form_tags"].form_tag = False
         context["footnotes"] = self.get_footnotes(context["measure"])
 
         conditions_initial = []
-
         if self.request.POST:
             conditions_initial = unprefix_formset_data(
                 MEASURE_CONDITIONS_FORMSET_PREFIX,
@@ -994,12 +994,12 @@ class MeasureUpdateBase(
             conditions_formset = forms.MeasureConditionsFormSet(
                 self.request.POST,
                 initial=conditions_initial,
-                prefix="measure-conditions-formset",
+                prefix=MEASURE_CONDITIONS_FORMSET_PREFIX,
             )
         else:
             conditions_formset = forms.MeasureConditionsFormSet(
                 initial=conditions_initial,
-                prefix="measure-conditions-formset",
+                prefix=MEASURE_CONDITIONS_FORMSET_PREFIX,
             )
         conditions = self.get_conditions(context["measure"])
         form_fields = conditions_formset.form.Meta.fields
